source/docs/mcp/001_function_calling_weather/pyproject.toml
[project]
name = "01_function_calling_weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.101.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
